Use float to form DAC output wave and calculate ADC input compensation parameter
 Note:
  Float is good enough precision to calculate wave data.
  The resolution of DAC(DAC102S085CIMM) is 10 bits.
  The resolution of ADC(ADCHS on LPC4370) is 12 bits.
  Using float makes object binary smaller, Cortex-M4 on
  LPC4370 has single precision (float) FPU.

+++ fw/program/include/calibrate.h
@@ -82,6 +82,6 @@ cmd_status_t calibrate_StoreCalibrationData(const calib_result_t * const data);
 cmd_status_t calibrate_EraseCalibrationData(void);
 const calib_result_t* calibrate_GetActiveCalibrationData(void);
 
-void calibrate_GetFactorsForDAC(int ch, double* a, double* b);
+void calibrate_GetFactorsForDAC(int ch, float* a, float* b);
 #endif /* end __CALIBRATE_H */
